requirements.txt
# Dashboard
streamlit>=1.35.0
plotly>=5.18.0
watchdog>=4.0.0

# Core
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Scraping
requests>=2.31.0
curl_cffi>=0.7.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0
httpx>=0.25.0
lxml>=4.9.0

# Database
SQLAlchemy>=2.0.0

# DFS Optimization
pulp>=2.7.0

# Scheduling / Automation
schedule>=1.2.0
APScheduler>=3.10.0
python-dotenv>=1.0.0

# CLI & Display
rich>=13.0.0
click>=8.1.0
tabulate>=0.9.0

# Charting / Audit Reports
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.18.0

# DataGolf (future — API wrapper placeholder)
# datagolf-python  # uncomment when subscribed

# AI / Claude
anthropic>=0.39.0

# Utilities
pytz>=2023.3
python-dateutil>=2.8.0
tqdm>=4.66.0
